WIP for new block handling

+++ gomint-server/src/main/java/io/gomint/server/util/Things.java
@@ -15,7 +15,7 @@ import io.gomint.world.block.BlockFace;
  */
 public class Things {
 
-    public static BlockFace convertFromDataToBlockFace( byte data ) {
+    public static BlockFace convertFromDataToBlockFace( short data ) {
         switch ( data ) {
             case 0:
                 return BlockFace.DOWN;
